[CE-4174]: fix percentile computation (#21830)
* [CE-4174]: fix percentile computation

forgot to divide percentile value to fraction (p/100)

* [CE-4174]: fix unitTests

+++ 280-batch-processing/src/main/java/io/harness/batch/processing/config/k8s/recommendation/estimators/ContainerResourceRequirementEstimators.java
@@ -59,8 +59,7 @@ public class ContainerResourceRequirementEstimators {
   }
 
   public static ContainerResourceRequirementEstimator customRecommender(
-      Map<String, Long> minResources, int percentile) {
-    // TODO(UTSAV): verify SafetyMargin and ConfidenceMultiplier here
+      Map<String, Long> minResources, double percentile) {
     final ResourceEstimator requiredEstimator = PercentileEstimator.of(percentile, percentile);
     return cs -> {
       Map<String, String> resources =
